refactor ask app (branch coverage 97%)
