+++ PlaskFront/src/app/answer/answer.ts
@@ -3,7 +3,6 @@ export class Answer{
 	id: number;
     author: string;
     content: string;
-    question_id: number; 
-    isMine: boolean;
+    question_id: number;
     
 }
